Commit - 202209261046
(DoublyLinkedList.cpp) Completed Test Case 1 by adding a temp node that acts as the iterator as we loop through each element in the doubly linked list. As we loop through each element we print out the memory address of the node, the previous nodes memory address, the value stored in the node, and the memory address to the next node in the linked list. The console output has formatting included to properly display the null pointers with a width of 14 characters using the iomanip library.

Added two supporting screenshots that will show the structure of the doubly linked list and the memory locations as you traverse the linked list.