Commit - 202209281216
(DoublyLinkedList.cpp) Added the logic to determine if the set is a palindrome. In this function we first designate two temporary nodes to represent the head and tail. Then, we verify the set is not empty. Which, would be considered a palindrome. Subsequently, we compare the values from the temporary head and tail for equality. If they are the same, we proceed by moving the head forward and the tail backward until all elements have been compared. If the loop completes with all values being equal, the resulting set is a palindrome.